moved batchdir, batchfile, batchstr functions to a newly created file batch_download.py
